Newton solver for final size.

Newton solver for final size.

Usage

.solve_newton(
  contact_matrix,
  demography_vector,
  susceptibility,
  iterations = 10000L,
  tolerance = 1e-06
)

Arguments

contact_matrix

Social contact matrix. Entry \(mm_{ij}\) gives average number of contacts in group \(i\) reported by participants in group \(j\).

demography_vector

Demography vector. Entry \(pp_{i}\) gives proportion of total population in group \(i\) (model will normalise if needed).

susceptibility

A matrix giving the susceptibility of individuals in demographic group \(i\) and risk group \(j\).

iterations

Number of solver iterations. Defaults to 10,000.

tolerance

Solver error tolerance. Solving for final size ends when the error drops below this tolerance. Defaults to set 1e-6. Larger tolerance values are likely to lead to inaccurate final size estimates.

Value

A two dimensional array of final sizes per age-risk group.