finalsize: Calculate the Final Size of an Epidemic

Calculate the final size of a susceptible-infectious-recovered epidemic in a population with demographic variation in contact patterns and susceptibility to disease, as discussed in Miller (2012) doi:10.1007/s11538-012-9749-6 .

See also

Author

Maintainer: Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [copyright holder]

Authors:

Edwin Van Leeuwen edwin.vanleeuwen@ukhsa.gov.uk (ORCID) [copyright holder]

Adam Kucharski adam.kucharski@lshtm.ac.uk (ORCID) [copyright holder]

Other contributors:

Rosalind Eggo r.eggo@lshtm.ac.uk (ORCID) [contributor]

Hugo Gruson hugo.gruson@data.org (ORCID) [contributor]

Thibaut Jombart thibaut@data.org (ORCID) [contributor]

Andree Valle-Campos andree.valle-campos@lshtm.ac.uk (ORCID) [contributor]

Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [reviewer]